requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
SQLAlchemy==2.0.0
pytest==7.2.1
alembic==1.9.3
coverage==7.1.0
PyJWT==2.6.0
pytest==7.2.1
httpx==0.23.3
flake8==6.0.0